microservices/qz-workflow-orchestrator/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
google-cloud-firestore>=2.16.0
